requirements.txt
pypdf
pdf2image
pytesseract
pillow
pandas
streamlit
